requirements.txt
streamlit
streamlit-drawable-canvas
git+https://github.com/magnushelliesen/matrix-mapper.git
git+https://github.com/magnushelliesen/neural-network.git
google-cloud-storage
